Split inject.js
